Convert outdent action-creator to reducer.
