+++ src/action-creators/index.js
@@ -23,7 +23,6 @@ export { default as loadResource } from './loadResource'
 export { default as moveThoughtDown } from './moveThoughtDown'
 export { default as moveThoughtUp } from './moveThoughtUp'
 export { default as newThoughtAtCursor } from './newThoughtAtCursor'
-export { default as outdent } from './outdent'
 export { default as preloadSources } from './preloadSources'
 export { default as prependRevision } from './prependRevision'
 export { scaleFontDown, scaleFontUp } from './scaleSize'
