+++ src/reducers/index.js
@@ -23,6 +23,7 @@ import modalComplete from './modalComplete'
 import modalRemindMeLater from './modalRemindMeLater'
 import newThought from './newThought'
 import newThoughtSubmit from './newThoughtSubmit'
+import outdent from './outdent'
 import render from './render'
 import search from './search'
 import searchLimit from './searchLimit'
@@ -77,6 +78,7 @@ const reducerMap = {
   modalRemindMeLater,
   newThought,
   newThoughtSubmit,
+  outdent,
   prioritizeScroll,
   render,
   search,
