+++ src/shortcuts/outdent.js
@@ -5,9 +5,6 @@ import {
   isDocumentEditable,
 } from '../util'
 
-// action-creators
-import outdent from '../action-creators/outdent'
-
 // eslint-disable-next-line jsdoc/require-jsdoc
 const Icon = ({ fill = 'black', size = 20, style }) => <svg version="1.1" className="icon" xmlns="http://www.w3.org/2000/svg" width={size} height={size} fill={fill} style={style} viewBox="0 0 64 64" enableBackground="new 0 0 64 64">
   <path d="m54 8h-44c-1.104 0-2 .896-2 2s.896 2 2 2h44c1.104 0 2-.896 2-2s-.896-2-2-2z" />
@@ -30,6 +27,6 @@ export default {
 
     if (cursor.length < 2) return
 
-    dispatch(outdent())
+    dispatch({ type: 'outdent' })
   }
 }
